+++ v3/openapi/document.go
@@ -37,11 +37,14 @@ type Components struct {
 }
 
 type PathItem struct {
-    Get    *Operation `json:"get,omitempty"`
-    Post   *Operation `json:"post,omitempty"`
-    Put    *Operation `json:"put,omitempty"`
-    Patch  *Operation `json:"patch,omitempty"`
-    Delete *Operation `json:"delete,omitempty"`
+    Get     *Operation `json:"get,omitempty"`
+    Post    *Operation `json:"post,omitempty"`
+    Put     *Operation `json:"put,omitempty"`
+    Patch   *Operation `json:"patch,omitempty"`
+    Delete  *Operation `json:"delete,omitempty"`
+    Options *Operation `json:"options,omitempty"`
+    Head    *Operation `json:"head,omitempty"`
+    Trace   *Operation `json:"trace,omitempty"`
 }
 
 type Operation struct {